requirements.txt
tensorflow
opencv-python
numpy
pandas
matplotlib
scikit-learn
streamlit
pillow